moved compoundable out of concerns, as it is not a concern
