+++ lib/cable_ready/compoundable.rb
